solve bug in header file
