+++ include/SSDPErrorHandle.h
@@ -1,4 +1,4 @@
-#ifndef UDPERRORHANDLER_H
-#define UDPERRORHANDLER_H
+#ifndef SDPERRORHANDLER_H
+#define SDPERRORHANDLER_H
 #include "../src/SSDPErrorHandle.c"
 #endif
