+++ include/customDataTypes.h
@@ -1,4 +1,3 @@
-#include <gtk/gtk.h>
 #ifndef CUSTOMDATATYPE_H
 #define CUSTOMDATATYPE_H
 
@@ -12,22 +11,4 @@ struct ssdpMessage {
   int size;
   char message[1024];
 };
-
-typedef struct {
-  GtkWidget *label;
-  GtkWidget *button;
-} NetworkStatusData;
-
-typedef struct {
-  GtkWidget *h_box;
-  GtkWidget *v_box;
-  struct ssdpMessage result;
-} UiUpdateData;
-
-typedef struct {
-  GtkWidget *window;
-  guint timeout_id;
-  NetworkStatusData *free_data;
-} StartData;
-
 #endif
