+++ include/uiCustomData.h
@@ -1,0 +1,4 @@
+#ifndef UICUSTOMDATA_H
+#define UICUSTOMDATA_H
+#include "../src/ui/uiCustomData.c"
+#endif
