+++ src/main.c
@@ -1,3 +1,4 @@
+#include "../include/UDPErrorHandle.h"
 #include "activate.h"
 
 int main(int argc, char **argv) {
